+++ nmos-web-patcher.py
@@ -2,14 +2,29 @@ from flask import Flask
 from routes.main import main_bp
 from routes.settings import settings_bp
 from routes.api import api_bp
-from services.cache import refresh_discovery, start_auto_refresh
 from __version__ import __version__
+import json
 import os
 
 app = Flask(__name__)
 app.register_blueprint(main_bp)
 app.register_blueprint(settings_bp)
 app.register_blueprint(api_bp)
+
+def load_settings():
+    try:
+        with open("settings.json") as f:
+            return json.load(f)
+    except:
+        return {}
+
+settings = load_settings()
+if settings.get("enable_restapi", False):
+    from protocols.restapi import restapi_bp
+    app.register_blueprint(restapi_bp)
+    print("[SETTINGS] REST API Enabled")
+else:
+    print("[SETTINGS] REST API Disabled")
 
 @app.context_processor
 def inject_version():
@@ -25,6 +40,6 @@ if __name__ == "__main__":
     print(f"[CREDITS] NMOS Web Patcher v{__version__} starting...")
     print(f"[CREDITS] by Arnaud Cresp")
     print(f"[CREDITS] https://github.com/acresp/nmos-web-patcher")
-    print(f"[CREDITS] You find this tool usefull? Buy me a coffee :)")
     print(f"[CREDITS] https://coff.ee/acresp")
+
     app.run(host="0.0.0.0", port=5000, debug=True)